backend/requirements.txt
torch>=1.13.1
torchvision>=0.14.1
albumentations>=1.3.0
opencv-python>=4.5.5.62
scikit-learn==1.2.2
matplotlib>=3.5.3
seaborn>=0.11.2
tqdm>=4.64.1
numpy>=1.21.6
pandas>=1.3.5
Pillow>=8.4.0
